have used a single instance of AVPlayer for the whole controller
Using a single controller for the AVPlayer is good for memory consumption but it creates problem when scrolling as the last frame of previous video is shown, need to work on it. The caching logic has no limit check on itself, the cache size could very well scale to a large number. need to implement logic to limit that